file restructure, help command now loops through table
