+++ uart_shell_arduino/src/cli/cli.h
@@ -1,0 +1,10 @@
+#ifndef CLI_H
+#define CLI_H
+
+// cli.h
+// Initializes UART CLI — also calls Serial.begin()
+void cli_init();
+
+// Handles UART input and command dispatch — call this in loop()
+void cli_poll();
+#endif
