+++ uart_shell_arduino/src/cli/commands.h
@@ -1,0 +1,23 @@
+#ifndef COMMANDS_H
+#define COMMANDS_H
+
+typedef void (*CommandFunction)(char* args);
+
+struct CommandEntry {
+  const char* name;               // command name
+  CommandFunction function;       // function to call
+  const char* description;        // command description
+};
+
+// expose command table
+extern CommandEntry commands[];
+
+// Function prototypes
+void handleCommand(char* cmd);
+void cmd_help(char* args);
+void cmd_status(char* args);
+void cmd_version(char* args);
+void cmd_say_hello(char* args);
+
+
+#endif
